Moves account-related listing logic to service layer
Transfers responsibility for listing entries by account from the account service to the entry service, promoting clearer separation of business concerns and improving maintainability.
