+++ src/main/java/me/josecomparotto/contabilidade_pessoal/controller/api/ContaApiController.java
@@ -19,6 +19,7 @@ import me.josecomparotto.contabilidade_pessoal.model.dto.conta.ContaViewDto;
 import me.josecomparotto.contabilidade_pessoal.model.dto.lancamento.LancamentoPartidaDto;
 import me.josecomparotto.contabilidade_pessoal.model.dto.conta.ContaNewDto;
 import me.josecomparotto.contabilidade_pessoal.service.ContaService;
+import me.josecomparotto.contabilidade_pessoal.service.LancamentoService;
 
 @RestController
 @RequestMapping("/api/contas")
@@ -26,6 +27,9 @@ public class ContaApiController {
 
     @Autowired
     private ContaService contasService;
+
+    @Autowired
+    private LancamentoService lancamentoService;
 
     // GET /api/contas
     @GetMapping
@@ -106,7 +110,7 @@ public class ContaApiController {
     // GET /api/contas/{id}/lancamentos
     @GetMapping("/{id}/lancamentos")
     public ResponseEntity<?> listarLancamentosPorConta(@PathVariable Integer id) {
-        List<LancamentoPartidaDto> lancamentos = contasService.listarLancamentosPorConta(id);
+        List<LancamentoPartidaDto> lancamentos = lancamentoService.listarLancamentosPorConta(id);
         if (lancamentos == null || lancamentos.isEmpty()) {
             return ResponseEntity.notFound().build();
         }
